* Implement click functionality for `ListItem`
