+++ src/App/index.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 
-import SideBar from "../components/SideBar";
+import SideBar from "../components/Sidebar";
 
 import useAxios from "../common/hooks/useAxios";
