+++ src/components/CatLoader.tsx
