--- src/components/List/styles.css
@@ -1,3 +1,0 @@
-div.list-items.active {
-  @apply fixed sm:relative h-full pb-24;
-}
