+++ src/components/styles.css
@@ -1,0 +1,16 @@
+nav.active {
+  @apply w-full sm:w-60 fixed sm:relative sm:h-screen h-32;
+}
+
+div.sidebar-content.active,
+.title.active {
+  @apply opacity-100;
+}
+
+div.list-items.active {
+  @apply fixed sm:relative h-full pb-24;
+}
+
+div.sidebar-list-item.active {
+  @apply sm:bg-orange-200;
+}
