+++ src/routes/BreedPage.tsx
@@ -1,0 +1,15 @@
+const BreedPage = () => {
+  return (
+    <div className="flex w-full h-screen p-4">
+      <div className="m-auto text-center h-4/6 space-y-4">
+        <h1 className="sm:text-3xl text-2xl font-pacifico">Breed Page</h1>
+        <p className="font-open-sans">
+          To get started, please select a cat breed to know more about its
+          information.
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default BreedPage;
