+++ src/routes/CatDetails.tsx
