--- src/routes/Dashboard/index.tsx
@@ -1,5 +1,0 @@
-const Dashboard = () => {
-  return <div className="w-full">Dashboard UI</div>;
-};
-
-export default Dashboard;
