+++ src/routes/index.tsx
@@ -7,11 +7,13 @@ import {
 import App from "../App";
 import CatDetails from "./CatDetails";
 import Dashboard from "./Dashboard";
+import BreedPage from "./BreedPage";
 
 const router = createBrowserRouter(
   createRoutesFromElements(
     <Route path="/" element={<App />}>
       <Route path="" element={<Dashboard />} />
+      <Route path="breed/:id" element={<BreedPage />} />
       <Route path="cat-details/:id" element={<CatDetails />} />
     </Route>
   )
